backend/requirements.txt
fastapi
uvicorn
openai
python-dotenv
azure-search-documents
azure-core
azure-identity
pydantic

# Database
sqlmodel
psycopg2-binary
alembic
